--- RestWithASPNETUdemy/RestWithASPNET8Udemy/Model/Context/MySQLContext.cs
@@ -1,13 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-
-namespace RestWithASPNETUdemy.Model.Context
-{
-    public class MySQLContext : DbContext
-    {
-        public MySQLContext() { }
-
-        public MySQLContext(DbContextOptions<MySQLContext> options) : base(options) { }
-
-        public DbSet<Person> Persons { get; set; }
-    }
-}
