--- RestWithASPNETUdemy/RestWithASPNET8Udemy/Model/Person.cs
@@ -1,22 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations.Schema;
-
-namespace RestWithASPNETUdemy.Model
-{
-    [Table("person")]
-    public class Person
-    {
-        [Column("id")]
-        public long Id { get; set; }
-        [Column("first_name")]
-        public string FirstName { get; set; }
-
-        [Column("last_name")]
-        public string LastName { get; set; }
-
-        [Column("address")]
-        public string Address { get; set; }
-
-        [Column("gender")]
-        public string Gender { get; set; }
-    }
-}
