--- RestWithASPNETUdemy/RestWithASPNET8Udemy/RestWithASPNET8Udemy.http
@@ -1,6 +1,0 @@
-@RestWithASPNET8Udemy_HostAddress = http://localhost:5226
-
-GET {{RestWithASPNET8Udemy_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
