--- RestWithASPNETUdemy/RestWithASPNET8Udemy/Services/IPersonService.cs
@@ -1,14 +1,0 @@
-﻿using RestWithASPNETUdemy.Model;
-using System.Collections.Generic;
-
-namespace RestWithASPNETUdemy.Services
-{
-    public interface IPersonService
-    {
-        Person Create(Person person);
-        Person FindById(long id);
-        List<Person> FindAll();
-        Person Update(Person person);
-        void Delete(long id);
-    }
-}
