--- RestWithASPNETUdemy/RestWithASPNET8Udemy/Services/Implementations/PersonServiceImplementation.cs
@@ -1,73 +1,0 @@
-﻿using RestWithASPNETUdemy.Model;
-using RestWithASPNETUdemy.Model.Context;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace RestWithASPNETUdemy.Services.Implementations
-{
-    public class PersonServiceImplementation : IPersonService
-    {
-        private MySQLContext _context;
-
-        public PersonServiceImplementation(MySQLContext context)
-        {
-            _context = context;
-        }
-
-        public Person Create(Person person)
-        {
-            try
-            {
-                _context.Add(person);
-                _context.SaveChanges();
-            }
-            catch (Exception ex)
-            {
-                throw;
-            }
-
-            return person;
-        }
-
-        public void Delete(long id)
-        {
-        }
-
-        public List<Person> FindAll()
-        {
-            return _context.Persons.ToList();
-        }
-
-        public Person FindById(long id)
-        {
-            return _context.Persons.SingleOrDefault(p => p.Id.Equals(id));
-        }
-
-        public Person Update(Person person)
-        {
-            if (!Exists(person.Id)) return new Person();
-
-            var result = _context.Persons.SingleOrDefault(p => p.Id.Equals(person.Id));
-
-            if (result != null)
-            {
-                try
-                {
-                    _context.Entry(result).CurrentValues.SetValues(person);
-                    _context.SaveChanges();
-                }
-                catch (Exception ex)
-                {
-                    throw;
-                }
-            }
-            return person;
-        }
-
-        private bool Exists(long id)
-        {
-            return _context.Persons.Any(p => p.Id.Equals(id));
-        }
-    }
-}
