--- RestWithASPNETUdemy/RestWithASPNET8Udemy/appsettings.json
@@ -1,12 +1,0 @@
-{
-  "MySQLConnection": {
-    "MySQLConnectionString": "Server=localhost;DataBase=rest_with_aspnet_udemy;Uid=root;Pwd=@Ll302508"
-  },
-    "Logging": {
-      "LogLevel": {
-        "Default": "Information",
-        "Microsoft.AspNetCore": "Warning"
-      }
-    },
-    "AllowedHosts": "*"
-  }
