+++ RestWithASPNETUdemy/RestWithASPNETUdemy/appsettings.Development.json
@@ -2,8 +2,7 @@
   "Logging": {
     "LogLevel": {
       "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
+      "Microsoft.AspNetCore": "Warning"
     }
   }
 }
