+++ RestWithASPNETUdemy/RestWithASPNETUdemy/appsettings.json
@@ -1,13 +1,12 @@
 {
   "MySQLConnection": {
     "MySQLConnectionString": "Server=localhost;DataBase=rest_with_aspnet_udemy;Uid=root;Pwd=@Ll302508"
-  }, 
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
   },
-  "AllowedHosts": "*"
-}
+    "Logging": {
+      "LogLevel": {
+        "Default": "Information",
+        "Microsoft.AspNetCore": "Warning"
+      }
+    },
+    "AllowedHosts": "*"
+  }
